+++ vllm/transformers_utils/configs/__init__.py
@@ -27,6 +27,7 @@ _CLASS_TO_MODULE: dict[str, str] = {
     "DeepseekVLV2Config": "vllm.transformers_utils.configs.deepseek_vl2",
     "DotsOCRConfig": "vllm.transformers_utils.configs.dotsocr",
     "EAGLEConfig": "vllm.transformers_utils.configs.eagle",
+    "FireRedASRConfig": "vllm.transformers_utils.configs.fireredasr",
     "FlexOlmoConfig": "vllm.transformers_utils.configs.flex_olmo",
     "FunAudioChatConfig": "vllm.transformers_utils.configs.funaudiochat",
     "FunAudioChatAudioEncoderConfig": "vllm.transformers_utils.configs.funaudiochat",
@@ -86,6 +87,7 @@ __all__ = [
     "DeepseekV3Config",
     "DotsOCRConfig",
     "EAGLEConfig",
+    "FireRedASRConfig",
     "FlexOlmoConfig",
     "FunAudioChatConfig",
     "FunAudioChatAudioEncoderConfig",
